Split ThumbLayoutManager to separate classes with shared base
