+++ DLCS.Model/DLCS.Model.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="iiif-net" Version="0.1.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Assets\Thumbs" />
+  </ItemGroup>
+
 </Project>
