+++ DLCS.Repository.Tests/DLCS.Repository.Tests.csproj
@@ -33,7 +33,6 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="Collections" />
       <Folder Include="Strategy" />
     </ItemGroup>
