+++ Engine/Ingest/Image/ImageOnDisk.cs
